feat: change type of index offset; usize to i64

+++ mjolnir/input/read_external_interaction.hpp
@@ -182,7 +182,7 @@ read_external_position_restraint_interaction(const toml::value& external)
         for(auto para : parameters)
         {
             const auto idx = find_parameter<std::size_t>(para, env, "index") +
-                             find_parameter_or<std::size_t>(para, env, "offset", 0);
+                             find_parameter_or<std::int64_t>(para, env, "offset", 0);
             const auto pos = find_parameter<std::array<real_type, 3>>(para, env, "position");
             const auto crd = math::make_coordinate<coordinate_type>(pos[0], pos[1], pos[2]);
